Lembretes: coluna ordem vira bigint
O app numera lembrete por ordem de criacao com Date.now() (~1.79e12), mas a
coluna `lembretes.ordem` nasceu int (max 2.147.483.647), entao todo lembrete
novo era recusado com "value ... is out of range for type integer". A agenda usa
o mesmo `ordem: Date.now()` e escapou porque la esse numero mora so dentro do
jsonb — lembretes e a unica tabela que copia o ordem para uma coluna solta.

migration_lembretes_ordem_bigint.sql conserta o banco que ja existe (ALTER
seguro com dados dentro e repetivel); migration_lembretes.sql passa a nascer
bigint para instalacao nova.

+++ supabase/migration_lembretes.sql
@@ -9,7 +9,7 @@ create table if not exists public.lembretes (
   texto      text,
   prazo      date,
   arquivado  boolean default false,
-  ordem      int,
+  ordem      bigint,
   updated_at timestamptz default now(),
   data       jsonb not null
 );
